% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\docType{methods}
\name{addError,funData-method}
\alias{addError,funData-method}
\title{Add gaussian white noise to functional data}
\usage{
\S4method{addError}{funData}(funDataObject, sd)
}
\description{
Add gaussian white noise to functional data
}
\keyword{internal}

