% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\docType{methods}
\name{addError,multiFunData-method}
\alias{addError,multiFunData-method}
\title{Add gaussian white noise to multivariate functional data}
\usage{
\S4method{addError}{multiFunData}(funDataObject, sd)
}
\description{
Add gaussian white noise to multivariate functional data
}
\keyword{internal}

