% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MFPCA_calculation.R
\name{calcMFPCA}
\alias{calcMFPCA}
\title{Internal function that implements the MFPCA algorithm for given univariate decompositions}
\usage{
calcMFPCA(N, p, Bchol, M, type, weights, npc, xVal, uniBasis, Yhat = FALSE,
  approx.eigen = TRUE)
}
\description{
Internal function that implements the MFPCA algorithm for given univariate decompositions
}
\keyword{internal}

