% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCP_TPA.R
\name{check_FCP_TPA_input}
\alias{check_FCP_TPA_input}
\title{Check input of FCP TPA function}
\usage{
check_FCP_TPA_input(X, K, penMat, alphaRange, verbose, tol, maxIter, adaptTol)
}
\value{
No return value, called for side effects
}
\description{
Check input of FCP TPA function
}
\keyword{internal}
