% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_mfpccox.R
\name{cv_mfpccox}
\alias{cv_mfpccox}
\title{Cross validation procedure for MFPCCox}
\usage{
cv_mfpccox(
  mFData,
  X_baseline,
  Y_surv,
  landmark_time = NULL,
  FakeLM = FALSE,
  times_pred = NULL,
  M = 50,
  uniExpansions = NULL,
  age = NULL,
  AgeDM = FALSE,
  type = c("scores", "AUC", "pp", "uscores"),
  n_folds = 10,
  seed = 1041996,
  verbose = FALSE,
  reg_baseline = FALSE,
  reg_long = TRUE,
  IPCW_vars = "none",
  de_bug = FALSE,
  truecdf = NULL
)
}
\arguments{
\item{X_baseline}{A data frame containing baseline covariates}

\item{Y_surv}{A 2 column data frame containing (time, status) survival info}

\item{landmark_time}{Time at which to landmark the data}

\item{FakeLM}{Should Fake (Relaxed) landmarking be performed? Default = False.}

\item{times_pred}{Times at which to predict the survival prob/AUC}

\item{M}{Number of dimensions used to approximate the process in mFPCA.}

\item{uniExpansions}{Univariate expansions to use in MFPCA, see MFPCA().}

\item{age}{Variable to perform Age Adjustment with.}

\item{AgeDM}{Should an extra step of Age Adjustment be performed? (Outdated, keep at FALSE)}

\item{type}{What measure should be used to summarize the longitudinal processes? Choice between mFPCA scores (scores), Area under Curve (AUC), 
predicted process (pp) and univariate scores (uscores). Default = scores.}

\item{n_folds}{Number of folds for cross validation.}

\item{seed}{Seed for random number generation}

\item{verbose}{Should progress messages be displayed?}

\item{reg_baseline}{Should baseline variables be LASSO regularized? Default = FALSE.}

\item{reg_long}{Should longitudinal variables be LASSO regularized? Default = TRUE.}

\item{IPCW_vars}{Vector indicating which variables should be used for Inverse Probability of Censoring Weights (IPCW) in determining validation scores.
Default is "none", meaning no IPCW will be performed.}

\item{de_bug}{Development option. Returns intermediate steps in the output. Default = FALSE.}

\item{truecdf}{A matrix indicating the true absolute risk (not landmarked) with column names indicating times and rows indicating subjects.}

\item{mdat}{A multiFunData object containing longitudinal covariates}
}
\description{
Cross validate MFPCCox procedure
}
\keyword{internal}
