% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/univDecomp.R
\name{dct3D}
\alias{dct3D}
\title{Calculate and threshold DCT for an 3D image}
\usage{
dct3D(image, qThresh)
}
\arguments{
\item{image}{A 3D image (a 3D array with real values).}

\item{qThresh}{A numeric with value in \eqn{[0,1]}, giving the quantile for
  thresholding the coefficients. See \code{\link{dctBasis3D}} for details.}
}
\value{
\item{ind}{An integer vector, containing the indices of
  non-thresholded (hence non-zero) coefficients.} \item{val}{A numeric
  vector, giving the values of the corresponding coefficients.}
}
\description{
This function calculates the (orthonormal) discrete cosine transformation for
a 3D image and returns thresholded DCT coefficients using the C-library
\code{fftw3} (see \url{http://www.fftw.org/}).
}
\section{Warning}{
 If the C-library \code{fftw3} is not available when the
  package \code{MFPCA} is installed, this function is disabled an will throw
  an error. For full functionality install the C-library \code{fftw3} from
  \url{http://www.fftw.org/} and reinstall \code{MFPCA}.
}
\seealso{
\code{\link{dctBasis3D}}
}
\keyword{internal}

