% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/univDecomp.R
\name{dctBasis3D}
\alias{dctBasis3D}
\title{Calculate a cosine basis representation for functional data on
three-dimensional domains}
\usage{
dctBasis3D(funDataObject, qThresh, parallel = FALSE)
}
\arguments{
\item{funDataObject}{An object of class \code{\link[funData]{funData}}
containing the observed functional data samples and for which the basis
representation is calculated.}

\item{qThresh}{A numeric with value in \eqn{[0,1]}, giving the quantile for
thresholding the coefficients. See Details.}

\item{parallel}{Logical. If \code{TRUE}, the coefficients for the basis
  functions are calculated in parallel. The implementation is based on the
  \code{\link[foreach]{foreach}} function and requires a parallel backend
  that must be registered before. See \code{\link[foreach]{foreach}} for
  details.}
}
\value{
\item{scores}{A sparse matrix of scores (coefficients) with dimension
  \code{N x L}, reflecting the weights \eqn{\theta_{mnk}} for each basis
  function in each observation, where \code{L} is the total number of basis
  functions used.} \item{B}{A diagonal matrix, giving the norms of the
  different basis functions used (as they are orthogonal).}
  \item{ortho}{Logical, set to \code{FALSE}, as basis functions are
  orthogonal, but in genereal not orthonormal.} \item{functions}{\code{NULL},
  as basis functions are known.}
}
\description{
This function calculates a tensor cosine basis representation for functional
data on three-dimensional domains based on a discrete cosine transformation
(DCT) using the C-library \code{fftw3} \url{http://www.fftw.org/}.
}
\details{
Given the (discretized) observed functions \eqn{X_i}, this function
calculates a basis representation \deqn{X_i(s,t, u) = \sum_{m = 0}^{M-1}
\sum_{n = 0}^{N-1} \sum_{k = 0}^{K-1} \theta_{mnk}  f_{mnk}(s,t,u)} in terms
of (orthogonal) tensor cosine basis functions \deqn{f_{mnk}(s,t,u) = c_m c_n
c_k \cos(ms) \cos(nt) \cos(ku), \quad (s,t,u) \in \mathcal{T}}{f_{mnk}(s,t,u) = c_m c_n
c_k \cos(ms) \cos(nt) \cos(ku), \quad (s,t,u) \in \calT} with \eqn{c_m =
\frac{1}{\sqrt{pi}}} for \eqn{m=0} and \eqn{c_m = \sqrt{\frac{2}{pi}}} for
\eqn{m=1,2,\ldots} based on a discrete cosine transform (DCT).

If not thresholded (\code{qThresh = 0}), the function returns all non-zero
coefficients \eqn{\theta_{mnk}} in the basis representation in a sparse
matrix \code{scores}. Otherwise, coefficients with \deqn{|\theta_{mnk}| <= q
} are set to zero, where \eqn{q} is the \code{qThresh}-quantile of
\eqn{|\theta_{mnk}|}.
}
\section{Warning}{
 If the C-library \code{fftw3} is not available when the
  package \code{MFPCA} is installed, this function is disabled an will throw
  an error. For full functionality install the C-library \code{fftw3} from
  \url{http://www.fftw.org/} and reinstall \code{MFPCA}.
}
\seealso{
univDecomp
}

