% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/univariateExpansions.R
\name{dctFunction3D}
\alias{dctFunction3D}
\title{Calculate linear combinations of orthonormal cosine basis functions on
three-dimensional domains}
\usage{
dctFunction3D(scores, xVal, parallel = FALSE)
}
\arguments{
\item{scores}{A sparse matrix of dimension \eqn{N x L}, representing the
\eqn{L} scores (coefficients) for each observation \eqn{i = 1, \ldots, N}.}

\item{xVal}{A list containing a two numeric vectors, corresponding to the x-
and y-values.}

\item{parallel}{Logical. If \code{TRUE}, the coefficients for the basis
  functions are calculated in parallel. The implementation is based on the
  \code{\link[foreach]{foreach}} function and requires a parallel backend
  that must be registered before. See \code{\link[foreach]{foreach}} for
  details.}
}
\value{
An object of class \code{funData} with \eqn{N} observations on the
  two-dimensional domain specified by \code{xVal}, corresponding to the
  linear combination of orthonormal cosine basis functions.
}
\description{
Given scores (coefficients), this function calculates a linear combination of
threse-dimensional cosine tensor basis functions on three-dimensional domains
using the C-library \code{fftw3} (see \url{http://www.fftw.org/}).
}
\section{Warning}{
 If the C-library \code{fftw3} is not available when the
  package \code{MFPCA} is installed, this function is disabled an will throw
  an error. For full functionality install the C-library \code{fftw3} from
  \url{http://www.fftw.org/} and reinstall \code{MFPCA}.
}
\seealso{
univExpansion
}

