% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/univariateExpansions.R
\name{defaultFunction}
\alias{defaultFunction}
\title{Calculate a linear combination of arbitrary basis function}
\usage{
defaultFunction(scores, xVal, functions)
}
\arguments{
\item{scores}{A matrix of dimension \eqn{N x K}, representing the \eqn{K}
scores (coefficients) for each observation \eqn{i = 1, \ldots, N}.}

\item{xVal}{A list containing a vector of x-values, representing a  \eqn{M_1
x ...x M_p} dimensional interval.}

\item{functions}{A \code{funData} object, representing \eqn{K} basis
  functions on a \eqn{M_1 x ...x M_p} dimensional domain.}
}
\value{
An object of class \code{funData} with \eqn{N} observations on
  \code{xVal}, corresponding to the linear combination of the basis
  functions.
}
\description{
This function calculates a linear combination of arbitrary basis funtions on
domains with arbitrary dimension.
}
\seealso{
univExpansion
}

