% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMean_Var.R
\name{DeMean_Var}
\alias{DeMean_Var}
\title{De-Mean data according to specified variable}
\usage{
DeMean_Var(mFData, var = NULL, nbasis = NULL, extraInfo = TRUE)
}
\arguments{
\item{mFData}{A multiFunData object}
}
\description{
Given realizations from a square integrable process, make the mean
over the realizations zero adjusted for a certain covariate.
}
\details{
Currently only numeric variables are supported
}
