% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{efPoly}
\alias{efPoly}
\title{Legendre Polynomials of degree 0,...,M-1}
\usage{
efPoly(argvals, M)
}
\arguments{
\item{argvals}{A vector, defining a (fine) grid on the interval for which the
Legendre polynomials are computed.}

\item{M}{An integer, specifying the number (and hence the degree) of
  polynomials that are calculcated.}
}
\value{
A univariate functional data object of class \code{\link[funData]{funData}}
  containing the Legendre polynomials on the given interval.
}
\description{
This function iteratively calculates orthonormal Legendre polynomials of
degree 0,...,M-1 on an arbitrary interval.
}
\seealso{
\code{\link[funData]{funData}}, \code{\link{simFunData}}, \code{\link{simMultiFunData}}.
}
\keyword{internal}

