% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{efWiener}
\alias{efWiener}
\title{Calculate the first M eigenfunctions of the Wiener process}
\usage{
efWiener(argvals, M)
}
\arguments{
\item{argvals}{A vector, defining a (fine) grid on the interval for which the
eigenfunctions are computed.}

\item{M}{An integer, specifying the number of eigenfunctions that are
  calculcated.}
}
\value{
A univariate functional data object of class
  \code{\link[funData]{funData}} containing the eigenfunctions of the Wiener
  process on the given interval.
}
\description{
This function calculates the first M (orthonormal) eigenfunctions of the
Wiener process on an arbitrary interval.
}
\seealso{
\code{\link[funData]{funData}}, \code{\link{simFunData}}, \code{\link{simMultiFunData}}.
}
\keyword{internal}

