% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{eVal}
\alias{eVal}
\title{Generate eigenvalues}
\usage{
eVal(M, type)
}
\arguments{
\item{M}{An integer, the number of eigenvalues to be generated.}

\item{type}{A character string specifying the type of eigenvalues that should
  be calculated. See Details.}
}
\value{
A vector containing the \eqn{M} decreasing eigenvalues.
}
\description{
This function generates \eqn{M} decreasing eigenvalues.
}
\details{
The function implements three types of eigenvalues: \itemize{\item
\code{"linear"} The eigenvalues start at  \eqn{1} and decrease linearly
towards \eqn{0}: \deqn{\nu_m = \frac{M+1 -m}{m}.}{\nu_m = (M+1-m)/m.} \item \code{"exponential"}
The eigenvalues start at \eqn{1}  and decrease exponentially towards \eqn{0}:
\deqn{\nu_m = \exp\left(-\frac{m-1}{2}\right).}{\nu_m = exp(-(m-1)/2).}\item \code{"wiener"} The eigenvalues
correspond to the eigenvalueso of the Wiener process: \deqn{\nu_m =
\frac{1}{(\pi/2 \cdot (2m-1))^2}.}{\nu_m = (pi/2 * (2m-1))^(-2)} }
}

