% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FCP_TPA.R
\name{FCP_TPA}
\alias{FCP_TPA}
\title{The functional CP-TPA algorithm}
\usage{
FCP_TPA(X, K, penMat, alphaRange, verbose = FALSE, tol = 1e-04,
  maxIter = 15, adaptTol = TRUE)
}
\arguments{
\item{X}{The data tensor of dimensions \code{N x S1 x S2}}

\item{K}{The number of eigentensors to be calculated}

\item{penMat}{A list of two roughness penalization matrices, one for each image
direction (v & w). No smoothing along observations.}

\item{verbose}{Logical. If \code{TRUE} computational details are given on the
 standard output during calculation of the FPCA}

\item{tol}{A numeric value, given the tolerance for relative and absolute values in the algorithm. Defaults to \code{1e-4}.}

\item{alphaVal}{A list of length 2, containing a vector of smoothing parameter values for each image direction (v & w).}
}
\value{
\item{d}{A vector of length \code{K}, containing the numeric weigths
  \eqn{d_k} in the CP model} \item{U}{A matrix of dimensions \code{N x K},
  containing the eigenvectors \eqn{u_k} in the first dimension} \item{V}{A
  matrix of dimensions \code{S1 x K}, containing the eigenvectors \eqn{v_k}
  in the second dimension} \item{W}{A matrix of dimensions \code{S2 x K},
  containing the eigenvectors \eqn{w_k} in the third dimension.}
}
\description{
This function implements the functional CP-TPA (FCP-TPA) algorithm, that
calculates a smooth PCA for 3D tensor data (i.e. \code{N} observations of 2D
images with dimension \code{S1 x S2}). The results are given in a
CANDECOMP/PARAFRAC (CP) model format \deqn{X = sum_{k = 1}^K d_k u_k o v_k o
w_k}  where \eqn{o} stands for the outer product. In this representation, the
outer product \eqn{v_k o w_k} can be regarded as the \eqn{k}-th eigenimage,
while \eqn{d_k u_k} represents the vector of individual scores for this
eigenimage and each observation.
}
\details{
The degree of smoothness is controlled by two smoothing parameters
\eqn{\alpha_v, \alpha_w}, each from a given grid in \code{alphaVal}, that are chosen via gcv
}
\references{
G. I. Allen, "Multi-way Functional Principal Components
  Analysis", IEEE International Workshop on Computational Advances in
  Multi-Sensor Adaptive Processing, 2013.
}
\seealso{
fcptpaBasis
}

