% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/univDecomp.R
\name{fcptpaBasis}
\alias{fcptpaBasis}
\title{Calculate an smooth PCA representaion for functional data on two-dimensional
domains}
\usage{
fcptpaBasis(funDataObject, npc, smoothingDegree = rep(2, 2), alphaRange)
}
\arguments{
\item{funDataObject}{An object of class \code{\link[funData]{funData}}
containing the observed functional data samples (here: images) for which
the smooth PCA is to be calculated.}

\item{npc}{An integer, giving a prespecified value for the number of
principal components.}

\item{smoothingDegree}{A numeric vector of length 2, specifying the degree of
the difference penalties inducing smoothness in both directions of the
image. Defaults to 2 for each direction (2nd differences).}

\item{alphaRange}{A list of length two with entries \code{v} and \code{w}
  containing the range of smoothness parameters to test for each direction.}
}
\value{
\item{scores}{A matrix of scores (coefficients) with dimension
  \code{N x k}, reflecting the weights for principal component in each
  observation.}  \item{B}{A matrix containing the scalar product of all pairs
  of basis functions.} \item{ortho}{Logical, set to \code{FALSE}, as basis
  functions are not orthonormal.} \item{functions}{A functional data object,
  representing the functional principal component basis functions.}
}
\description{
This function calculates a smooth PCA representation based on the FCP_TPA
algorithm for functional data on two-dimensional domains. In this case, the
data can be interpreted as images with \code{S1 x S2} pixels (assuming
\code{nObsPoints(funDataObject)} = (S1, S2)), i.e. the total observed data
are represented as third order tensor of dimension \code{N x S1 x S2}.  The
smooth PCA of a tensor of this kind is calculated via the
\code{\link{FCP_TPA}} function. The smoothness of the resulting eigenvectors
\eqn{u_k, v_k, w_k} (that are used for the calculation of the eigenimages and
the score vectors) is controlled by smoothing parameters \eqn{\alpha_u,
\alpha_v, \alpha_w}.
}
\references{
G. I. Allen, "Multi-way Functional Principal Components
  Analysis", In IEEE International Workshop on Computational Advances in
  Multi-Sensor Adaptive Processing, 2013.
}
\seealso{
univDecomp
}

