% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/univariateExpansions.R
\name{fcptpaFunction}
\alias{fcptpaFunction}
\title{Calculate a linear combination of a FCP_TPA basis on two-dimensional domains}
\usage{
fcptpaFunction(scores, argvals = functions@argvals, functions)
}
\arguments{
\item{scores}{A matrix of dimension \code{N x K}, representing the \code{K}
scores (coefficients) for each of the \code{N} observations.}

\item{argvals}{A list containing a vector of x-values. Defaults to
\code{functions@argvals}.}

\item{functions}{A \code{funData} object, representing the FCP_TPA basis.}
}
\value{
An object of class \code{funData} with \code{N} observations on
  \code{argvals}, corresponding to the linear combination of the functional
  principal components.
}
\description{
This function calculates a linear combination of basis functions based on
the FCP_TPA algorithm (see References) for data on
two-dimensional domains.
}
\examples{
# see expandBasisFunction
}
\references{
G. I. Allen, "Multi-way Functional Principal Components
  Analysis", In IEEE International Workshop on Computational Advances in
  Multi-Sensor Adaptive Processing, 2013.
}
\seealso{
\code{\link{univExpansion}}, \code{\link{expandBasisFunction}}
}

