% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MFPCA_calculation.R
\name{findUniArgs}
\alias{findUniArgs}
\title{Utility function for defining arguments of univariate basis expansion}
\usage{
findUniArgs(uniExpansion, funDataObject)
}
\arguments{
\item{uniExpansion}{A list corresponding to one univariate expansion as
described in \code{\link{MFPCA}}.}

\item{funDataObject}{An object of class \code{\link[funData]{funData}},
  representing the (univariate) functional data, for which the expansion will
  be calculated}
}
\value{
\item{args}{A list of parameters to be passed to
  \code{\link{univBasisExpansion}}}.
}
\description{
This is a utility function that defines the arguments
of\code{\link{univBasisExpansion}} in \code{\link{MFPCA}}. This is used as an
internal function in \code{\link{MFPCA}}.
}
\keyword{internal}

