% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mcox.R
\name{fit_mcox}
\alias{fit_mcox}
\title{Fit multivariate Cox model (outdated)}
\usage{
fit_mcox(
  mscores,
  Y_train,
  X_baseline_train = NULL,
  X_baseline_test = NULL,
  Y_test = NULL,
  landmark_time = NULL
)
}
\arguments{
\item{mscores}{output of predict_mscores function (contains scores for train and test data).}

\item{Y_train}{Survival information (time, event) for training data}

\item{X_baseline_train}{Baseline covariates data frame for training data}

\item{landmark_time}{Landmark time for predictions. See equation in Section 2.4 in Li/Luo}
}
\description{
Fit multivariate Cox model (outdated)
}
