% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/univDecomp.R
\name{fpcaBasis}
\alias{fpcaBasis}
\title{Calculate a functional principal component basis representation for
functional data on one-dimensional domains}
\usage{
fpcaBasis(funDataObject, nbasis, pve, npc, makePD)
}
\arguments{
\item{funDataObject}{An object of class \code{\link[funData]{funData}}
containing the observed functional data samples and for which the FPCA is
to be calculated.}

\item{nbasis}{An integer, representing the number of  B-spline basis
functions used for estimation of the mean function and bivariate smoothing
of the covariance surface. Defaults to 10 (cf.
\code{\link[refund]{fpca.sc}}).}

\item{pve}{A numeric value between 0 and 1, the proportion of variance
explained: used to choose the number of principal components. Defaults to
0.99 (cf. \code{\link[refund]{fpca.sc}}).}

\item{npc}{An integer, giving a prespecified value for the number of
principal components. Defaults to \code{NULL}. If given, this overrides
\code{pve} (cf. \code{\link[refund]{fpca.sc}}).}

\item{makePD}{Logical: should positive definiteness be enforced for the
  covariance surface estimate? Defaults to \code{FALSE} (cf.
  \code{\link[refund]{fpca.sc}}).}
}
\value{
\item{scores}{A matrix of scores (coefficients) with dimension
  \code{N x k}, reflecting the weights for principal component in each
  observation.} \item{B}{NULL, as functions are orthonormal.}
  \item{ortho}{Logical, set to \code{TRUE}, as basis functions are
  orthonormal.} \item{functions}{A functional data object, representing the
  functional principal component basis functions.}
}
\description{
This function calculates functional principal component basis representation
for functional data on one-dimensional domains. The FPCA is calculated via
the \link{PACE} function, which is built on \link[refund]{fpca.sc} in the
\pkg{refund} package.
}
\seealso{
univDecomp
}

