% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/univariateExpansions.R
\name{fpcaFunction}
\alias{fpcaFunction}
\title{Calculate a linear combination of a functional principal component basis on
one-dimensional domains}
\usage{
fpcaFunction(scores, argvals = functions@argvals, functions)
}
\arguments{
\item{scores}{A matrix of dimension \code{N x K}, representing the \code{K}
scores (coefficients) for each of the \code{N} observations.}

\item{argvals}{A list containing a vector of x-values. Defaults to
\code{functions@argvals}.}

\item{functions}{A \code{funData} object, representing the FPCA basis.}
}
\value{
An object of class \code{funData} with \code{N} observations on
  \code{argvals}, corresponding to the linear combination of the functional
  principal components.
}
\description{
This function calculates  a linear combination of functional principal
component basis functions on one-dimensional domains.
}
\examples{
# set seed
set.seed(1234)

# simulate coefficients (scores) for N = 10 observations and K = 8 basis functions
N <- 10
K <- 8
scores <- t(replicate(n = N, rnorm(K, sd = (K:1)/K)))
dim(scores)

# Fourier basis functions as eigenfunctions
eFuns <- eFun(argvals = seq(0,1,0.01), M = K, type = "Fourier")

# calculate PCA expansion
f <- fpcaFunction(scores = scores, functions = eFuns)

oldpar <- par(no.readonly = TRUE)

par(mfrow = c(1,2))
plot(eFuns, main = "Basis functions")
plot(f, main = "Linear combination")

par(oldpar)
}
\seealso{
\code{\link{univExpansion}}, \code{\link{fpcaBasis}}
}

