% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mscores.R
\name{get_mscores}
\alias{get_mscores}
\title{Predict multivariate scores for multiFunData (step 2)}
\usage{
get_mscores(step1, M = NULL, uniExpansions = NULL, verbose = FALSE)
}
\arguments{
\item{step1}{Output of landmark_data()}

\item{M}{Number of dimensions to use in mFPCA}

\item{uniExpansions}{Univariate expansions to use in mFPCA}

\item{verbose}{Should intermediary results be printed?}

\item{mFData_train}{Training data in multiFunData format. Can also supply an 
MFPCAfit already constructed on training data. In that case, uniExpansions 
does not have to be specified, but will instead be extracted from mFData_train.}

\item{mFData_pred}{Testing data in multiFunData format}

\item{age_train}{Age of patients at baseline for training data.}

\item{age_pred}{Age of patients at baseline for testing data}

\item{landmark_time}{Time at which to landmark}

\item{Y_train}{Survival info (time, event) for train data. Only required in case of landmarking!}
}
\description{
Time units in Y_train and mFData_train$argvals have to be the same, otherwise cannot do landmarking!!!
}
