% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmark_data.R
\name{landmark_data}
\alias{landmark_data}
\title{Landmark data (step 1)}
\usage{
landmark_data(
  time = NULL,
  mFData_train,
  X_baseline_train,
  Y_surv_train,
  age_train = NULL,
  mFData_pred = NULL,
  X_baseline_pred = NULL,
  Y_surv_pred = NULL,
  age_pred = NULL
)
}
\arguments{
\item{time}{Landmark time. Default = NULL (no landmarking).}

\item{mFData_train}{Training data in multiFunData format.}

\item{X_baseline_train}{Baseline covariates for training data}

\item{Y_surv_train}{Survival information (time, event) for training data}

\item{age_train}{(optional) Numeric vector of ages at baseline for training data}

\item{mFData_pred}{Data for which predictions are required in multiFunData format.}

\item{X_baseline_pred}{Baseline covariates for testing data}

\item{Y_surv_pred}{Survival information (time, event) for prediction data. Can be left empty,
in that case predictions will be made for the whole test data.}

\item{age_pred}{(optional) Numeric vector of ages at baseline for prediction data.}
}
\description{
Landmark data (remove information on patients with failures before landmark time).
}
\details{
mFData_train, X_baseline_train, Y_surv_train and age_train must have the same row order. Each row should represent the same patient.
The same is true for prediction data.
}
