% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMPCA.R
\name{maxeig}
\alias{maxeig}
\title{Compute the largest eigenvalue and associated eigenvector of a matrix A using the power method}
\usage{
maxeig(A)
}
\arguments{
\item{A}{Matrix whose eigenvalue is sought}
}
\value{
\item{lambda}{Largest eigenvalue}
\item{x}{corresponding eigenvector}
}
\description{
MATLAB Function: Copyright 1999 by Todd K. Moon
}
\keyword{internal}

