% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivExpansions.R
\name{multivExpansion}
\alias{multivExpansion}
\title{Calculate multivariate basis expansion}
\usage{
multivExpansion(multiFuns, scores)
}
\arguments{
\item{multiFuns}{A multivariate functional data object, containing the
multivariate basis functions}

\item{scores}{A matrix containing the scores for each observation in each
row. The number of columns must match the number of basis functions.}
}
\value{
A \code{multiFunData} object containing the expanded functions for
  each observation.
}
\description{
Calculate multivariate basis expansion
}
