% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FCP_TPA.R
\name{normVec}
\alias{normVec}
\title{Calculate the euclidean norm of a vector}
\usage{
normVec(x)
}
\arguments{
\item{x}{The vector for which the norm is to be calculated}
}
\value{
The euclidean norm of \code{x}.
}
\description{
Calculate the euclidean norm of a vector
}
\keyword{internal}

