% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.rcv_mfpccox}
\alias{plot.rcv_mfpccox}
\title{Plot performance measures of models}
\usage{
plot.rcv_mfpccox(x, ...)
}
\description{
Can plot output of cv and rcv
}
\section{Functions}{
\itemize{
\item \code{plot.rcv_mfpccox}: Plot a CGR-CUSUM
}}

