% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictRisk.R
\name{predict.GLMnet}
\alias{predict.GLMnet}
\title{Predict absolute risk}
\usage{
\method{predict}{GLMnet}(object, newdata, type = type, s, ...)
}
\arguments{
\item{object}{A 'cv.glmnet' object which was used to evaluate a 'coxnet' object.}

\item{newdata}{A data frame containing predictor variable combinations for which to compute predicted event probabilities}

\item{x}{Input matrix, see glmnet}

\item{y}{Response variable, see glmnet}

\item{times}{Times at which to predict the absolute risk}
}
\description{
Predict absolute risk for new data
}
