% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_surv.R
\name{predict_surv}
\alias{predict_surv}
\title{Predict survival probabilities for new patients (step 3)}
\usage{
predict_surv(step2, times_pred = step2$landmark_time)
}
\arguments{
\item{step2}{output of get_mscores}

\item{times_pred}{Times at which prediction is required. Default = landmark_time (useless).}
}
\description{
Predict survival for prediction data
}
