% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predsurv.R
\name{predsurv}
\alias{predsurv}
\title{Predict survival probabilities for new patients (step 3)}
\usage{
predsurv(
  step2,
  times_pred = step2$landmark_time,
  accuracy_train = FALSE,
  reg_baseline = FALSE,
  reg_long = TRUE
)
}
\arguments{
\item{step2}{output of get_mscores}

\item{times_pred}{Times at which prediction is required. Default = landmark_time (useless).}

\item{accuracy_train}{Should accuracy measures (AUC/Brier score) be calculated on the training set?}

\item{reg_baseline}{Should baseline covariates be regularized?}

\item{reg_long}{Should longitudinal covariates be regularized?}
}
\description{
Predict survival for prediction data
}
