% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFPCAfit_methods.R
\name{print.summary.MFPCAfit}
\alias{print.summary.MFPCAfit}
\title{Print summary of a Multivariate Functional Principal Component Analysis}
\usage{
\method{print}{summary.MFPCAfit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{MFPCAfit.summary}, usually returned by a
call to \link{MFPCAfit.summary}.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
A \code{print} method for class \code{MFPCAfit.summary}
}
