% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_mfpccox.R
\name{rcv_mfpccox}
\alias{rcv_mfpccox}
\title{Repeated Cross Validated Regularized AA MFPCACox procedure.}
\usage{
rcv_mfpccox(
  mFData,
  X_baseline,
  Y_surv,
  landmark_time = NULL,
  FakeLM = FALSE,
  times_pred = NULL,
  M = 50,
  uniExpansions = NULL,
  AgeDM = FALSE,
  age = NULL,
  type = c("scores", "AUC", "pp", "uscores"),
  n_reps = 10,
  n_folds = 10,
  seed = 1041996,
  displaypb = FALSE,
  n_cores = 1,
  verbose = FALSE,
  reg_baseline = FALSE,
  reg_long = TRUE,
  IPCW_vars = "none",
  trueSurv
)
}
\arguments{
\item{X_baseline}{A data frame containing baseline covariates}

\item{Y_surv}{A 2 column data frame containing (time, status) survival info}

\item{landmark_time}{Time at which to landmark the data}

\item{FakeLM}{Should Fake (Relaxed) landmarking be performed? Default = False.}

\item{times_pred}{Times at which to predict the survival prob/AUC}

\item{M}{Number of dimensions used to approximate the process in mFPCA.}

\item{uniExpansions}{Univariate expansions to use in MFPCA, see MFPCA().}

\item{AgeDM}{Should an extra step of Age Adjustment be performed? (Outdated, keep at FALSE)}

\item{age}{Variable to perform Age Adjustment with.}

\item{type}{What measure should be used to summarize the longitudinal processes? Choice between mFPCA scores (scores), Area under Curve (AUC), 
predicted process (pp) and univariate scores (uscores). Default = scores.}

\item{n_reps}{Number of repeats of the cross validation procedure}

\item{n_folds}{Number of folds for cross validation.}

\item{seed}{Seed for random number generation}

\item{displaypb}{Should a progress bar be displayed?}

\item{n_cores}{How many cores should be used?}

\item{verbose}{Should progress messages be displayed?}

\item{reg_baseline}{Should baseline variables be LASSO regularized? Default = FALSE.}

\item{reg_long}{Should longitudinal variables be LASSO regularized? Default = TRUE.}

\item{IPCW_vars}{Vector indicating which variables should be used for Inverse Probability of Censoring Weights (IPCW) in determining validation scores.
Default is "none", meaning no IPCW will be performed.}

\item{trueSurv}{A matrix indicating the true Survival probabilities. Allows for comparing predicted and true survival probability}
}
\description{
Same as cv_mfpccox(), but repeated instead of single cross validation.
}
\keyword{internal}
