% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFPCAfit_methods.R
\name{scoreplot}
\alias{scoreplot}
\title{Scoreplot Generic}
\usage{
scoreplot(PCAobject, ...)
}
\arguments{
\item{PCAobject}{A principal component object}

\item{...}{Arguments passed from or to other methods}
}
\value{
A bivariate plot of scores.
}
\description{
Redirects to \code{\link[graphics]{plot.default}}
}
