% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{simMultiSplit}
\alias{simMultiSplit}
\title{Simulate multivariate eigenfunctions based on a split 'big' ONB}
\usage{
simMultiSplit(argvals, M, eFunType, ignoreDeg = NULL, eValType, N)
}
\description{
Simulate multivariate eigenfunctions based on a split 'big' ONB
}
\keyword{internal}

