% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simMultiSplitv2}
\alias{simMultiSplitv2}
\title{Simulate multivariate eigenfunctions based on a split 'big' ONB}
\usage{
simMultiSplitv2(argvals, M, eFunType, ignoreDeg = NULL, eValType, N)
}
\description{
Simulate multivariate eigenfunctions based on a split 'big' ONB
}
\keyword{internal}
