% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{simMultiWeight}
\alias{simMultiWeight}
\title{Simulate multivariate eigenfunctions based on weighted orthonormal bases}
\usage{
simMultiWeight(argvals, M, eFunType, ignoreDeg = NULL, eValType, N)
}
\description{
Simulate multivariate eigenfunctions based on weighted orthonormal bases
}
\keyword{internal}

