% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simMultiWeightv2}
\alias{simMultiWeightv2}
\title{Simulate multivariate eigenfunctions based on weighted orthonormal bases}
\usage{
simMultiWeightv2(argvals, M, eFunType, ignoreDeg = NULL, eValType, N)
}
\description{
Simulate multivariate eigenfunctions based on weighted orthonormal bases
}
\keyword{internal}
