% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\docType{methods}
\name{sparsify,funData-method}
\alias{sparsify,funData-method}
\title{sparsify for univariate functional data}
\usage{
\S4method{sparsify}{funData}(funDataObject, minObs, maxObs)
}
\description{
sparsify for univariate functional data
}
\keyword{internal}

