% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\docType{methods}
\name{sparsify,multiFunData-method}
\alias{sparsify,multiFunData-method}
\title{sparsify for multivariate functional data}
\usage{
\S4method{sparsify}{multiFunData}(funDataObject, minObs, maxObs)
}
\description{
sparsify for multivariate functional data
}
\keyword{internal}

