% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/univariateExpansions.R
\name{splineBasis1Dpen}
\alias{splineBasis1Dpen}
\title{Calculate a penalized spline basis representation for functional data on
one-dimensional domains}
\usage{
splineBasis1Dpen(funDataObject, bs, m, k)
}
\arguments{
\item{funDataObject}{An object of class \code{\link[funData]{funData}}
containing the observed functional data samples and for which the bases
representation is calculated.}

\item{bs}{The type of basis functions to be used. Please refer to
\code{link[mgcv]{smooth.terms}} for a list of possible basis functions.}

\item{m}{The order of the spline basis. See  \code{link[mgcv]{s}} for
details.}

\item{k}{The number of basis functions used.  See  \code{link[mgcv]{s}} for
  details.}
}
\value{
scores A matrix of weights with dimension \code{N x k}, reflecting
  the weights for each basis function in each observation.

function A \code{\link[funData]{funData}} object representing the
  basis functions.
}
\description{
This function calculates a penalized spline basis representation for
functional data on one-dimensional domains based on the \link[mgcv]{gam}
function in the \pkg{mgcv} package.
}
\keyword{internal}

