% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/univariateExpansions.R
\name{splineFunction1D}
\alias{splineFunction1D}
\title{Calculate linear combinations of spline basis functions on one-dimensional
domains}
\usage{
splineFunction1D(scores, xVal, bs, m, k)
}
\arguments{
\item{scores}{A matrix of dimension \eqn{N x K}, representing the \eqn{K}
scores (coefficients) for each observation \eqn{i = 1, \ldots, N}.}

\item{xVal}{A list containing a vector of x-values.}

\item{bs}{A character string, specifying the type of basis functions to be
used. Please refer to \code{\link[mgcv]{smooth.terms}} for a list of
possible basis functions.}

\item{m}{A numeric, the order of the spline basis. See  \code{\link[mgcv]{s}}
for details.}

\item{k}{A numeric, the number of basis functions used.  See
  \code{\link[mgcv]{s}} for details.}
}
\value{
An object of class \code{funData} with \eqn{N} observations on
  \code{xVal}, corresponding to the linear combination of spline basis
  functions.
}
\description{
Given scores (coefficients), this function calculates a linear combination of
one-dimensional spline basis functions on one-dimensional domains based on
the \link[mgcv]{gam} function in the \pkg{mgcv} package.
}
\seealso{
univExpansion
}

