% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/univariateExpansions.R
\name{splineFunction2D}
\alias{splineFunction2D}
\alias{splineFunction2Dpen}
\title{Calculate linear combinations of spline basis functions on two-dimensional
domains}
\usage{
splineFunction2D(scores, xVal, bs, m, k)

splineFunction2Dpen(scores, xVal, bs, m, k)
}
\arguments{
\item{scores}{A matrix of dimension \eqn{N x K}, representing the \eqn{K}
scores (coefficients) for each observation \eqn{i = 1, \ldots, N}.}

\item{xVal}{A list containing a two numeric vectors, corresponding to the x-
and y-values.}

\item{bs}{An vector of character strings (or a single character), the type of
basis functions to be used. Please refer to \code{\link[mgcv]{te}} for a
list of possible basis functions.}

\item{m}{A numeric vector (or a single number), the order of the spline
basis. See \code{\link[mgcv]{s}} for details.}

\item{k}{A numeric vector (or a single number), the number of basis functions
  used.  See  \code{\link[mgcv]{s}} for details.}
}
\value{
An object of class \code{funData} with \eqn{N} observations on the
  tow-dimensional domain specified by \code{xVal}, corresponding to the
  linear combination of spline basis functions.
}
\description{
Given scores (coefficients), this function calculates a linear combination of
two-dimensional spline tensor basis functions on one-dimensional domains
based on the \link[mgcv]{gam} function in the \pkg{mgcv} package. If the
scores have been calculated based on a penalized tensor spline basis, use
\code{splineFunction2Dpen} instead (which runs \link[mgcv]{bam} instead of
\link[mgcv]{gam}).
}
\seealso{
univExpansion
}

