% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_validation.R
\name{true_validation}
\alias{true_validation}
\title{Calculate true validation scores if survival probability of subjects is known}
\usage{
true_validation(trueCDF, Y_surv, landmark_time = NULL)
}
\arguments{
\item{trueCDF}{CDF of observing an event at pred_time (matrix), with rows indicating subjects and columns the prediction times.
Column names should be the prediction times of interest. First column must be landmark time.}

\item{Y_surv}{Matrix of survival and censoring times (time, status)}

\item{landmark_time}{Landmark time for validation. Default = NULL will use first available time.}
}
\description{
Calculate true tdAUC and Brier score given inputs.
}
\keyword{internal}
