% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMPCA.R
\name{ttvCalculation}
\alias{ttvCalculation}
\title{Internal function for the Tensor times Vector calculation}
\usage{
ttvCalculation(A, v, dim)
}
\arguments{
\item{A}{Tensor}

\item{v}{Vector}

\item{dim}{Dimension}
}
\value{
Result of tensor times vector calculation
}
\description{
Internal function for the Tensor times Vector calculation
}
\seealso{
\code{\link{ttv}}
}
\keyword{internal}
