% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UMPCA.R
\name{UMPCA}
\alias{UMPCA}
\title{UMPCA: Uncorrelated Multilinear Principle Component Analysis}
\usage{
UMPCA(TX, numP)
}
\arguments{
\item{TX}{The input training data in tensorial representation, the last mode
is the sample mode. For Nth-order tensor data, \code{TX} is of
(N+1)th-order with the (N+1)-mode to be the sample mode.
E.g., 30x20x10x100 for 100 samples of size 30x20x10#'}

\item{numP}{The dimension of the projected vector, denoted as \eqn{P} in the
         paper. It is the number of elementary multilinear projections
         (EMPs) in tensor-to-vector projection.}
}
\value{
\item{Us}{The multilinear projection, consiting of \code{numP} (\eqn{P} in the paper)
       elementary multilinear projections (EMPs), each EMP is consisted
       of \eqn{N} vectors, one in each mode.}
 \item{TXmean}{The mean of the input training samples \code{TX}.}
 \item{odrIdx}{The ordering index of projected features in decreasing
           variance.}
}
\description{
This function implements the uncorrelated multilinear principal component
analysis for tensors of dimension 2, 3 or 4. The code is basically the same as
in the  MATLAB toolbox
\url{UMPCA}{http://www.mathworks.com/matlabcentral/fileexchange/35432-uncorrelated-multilinear-principal-component-analysis--umpca-}
by Haiping Lu (see references).
}
\references{
Haiping Lu, K.N. Plataniotis, and A.N. Venetsanopoulos,
  "Uncorrelated Multilinear Principal Component Analysis for Unsupervised Multilinear Subspace Learning",
   IEEE Transactions on Neural Networks,
  Vol. 20, No. 11, Page: 1820-1836, Nov. 2009.
}

