% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/univariateExpansions.R
\name{univExpansion}
\alias{univExpansion}
\title{Calculate a univariate basis expansion}
\usage{
univExpansion(type, scores, xVal, functions, params)
}
\arguments{
\item{type}{A character string, specifying the basis for which the
decomposition is to be calculated.}

\item{scores}{A matrix of scores (coefficients) for each observation based on
the given basis functions.}

\item{xVal}{A list, representing the domain of the basis functions. See
\linkS4class{funData} for details.}

\item{functions}{A functional data object, representing the basis functions.
Can be \code{NULL} if the basis functions are not estimated from observed
data, but have a predefined form. See Details.}

\item{params}{A list containing the parameters for the particular basis to
  use.}
}
\value{
A functional data object, representing the linear combinations of the
  basis functions based on the given scores.
}
\description{
This function calculates a univariate basis expansion based on given scores
(coefficients) and basis functions.
}
\details{
This function calculates functional data \eqn{X_i(t), i= 1 \ldots N} that is
represented as a linear combination of basis functions \eqn{b_k(t)}
\deqn{X_i(t) = \sum_{k = 1}^K \theta_{ik} b_k(t), i = 1, \ldots, N.} The
basis functions may be prespecified (such as spline basis functions or
Fourier bases) or can be estimated from observed data (e.g. by functional
principal component analysis)
}

