% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model.R
\name{validate_model}
\alias{validate_model}
\title{Validate prediction model (step 4)}
\usage{
validate_model(
  predRisk,
  preddat,
  times_pred,
  reg_baseline = FALSE,
  reg_long = TRUE,
  IPCW_vars = c("none", "all"),
  truecdf = NULL,
  landmark_time,
  FakeLM
)
}
\arguments{
\item{predRisk}{matrix containing the predicted cdf of subjects (rows) over times_pred (columns)}

\item{preddat}{data frame containing survival time, status indicator en covariates}

\item{times_pred}{Prediction times at which the models should be validated}

\item{reg_baseline}{Should baseline covariates be regularized?}

\item{reg_long}{Should longitudinal covariates be regularized?}

\item{IPCW_vars}{Character vector indicating which variables to use for IPCW.
Extra options are "all" (all baseline variables) and "none" (don't perform IPCW). Default is no IPCW ("none").}

\item{truecdf}{matrix containing the true cdf of subjects (rows) over times_pred (columns)}

\item{landmark_time}{Landmark time}

\item{FakeLM}{Are we performing Relaxed landmarking? Default = FALSE}
}
\description{
Validate prediction model by calculating tdAUC, Brier Score & MSE (if true cdf specified)
}
\keyword{internal}
