% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFPCA_calculation.R
\name{calcMFPCA}
\alias{calcMFPCA}
\title{Internal function that implements the MFPCA algorithm for given univariate decompositions}
\usage{
calcMFPCA(N, p, Bchol, M, type, weights, npc, argvals, uniBasis, fit = FALSE,
  approx.eigen = FALSE)
}
\description{
Internal function that implements the MFPCA algorithm for given univariate decompositions
}
\keyword{internal}
