% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_mfpccox.R
\name{cv_mfpccox}
\alias{cv_mfpccox}
\title{Cross validation procedure for MFPCCox}
\usage{
cv_mfpccox(
  mFData,
  X_baseline,
  Y_surv,
  landmark_time = 0,
  times_pred = NULL,
  M = 50,
  uniExpansions = NULL,
  age = NULL,
  AgeDM = FALSE,
  type = c("scores", "AUC", "pp", "uscores"),
  n_folds = 10,
  seed = 1041996,
  verbose = FALSE,
  reg_baseline = FALSE,
  reg_long = TRUE,
  IPCW_vars = NULL
)
}
\arguments{
\item{X_baseline}{A data frame containing baseline covariates}

\item{Y_surv}{A 2 column data frame containing (time, status) survival info}

\item{landmark_time}{Time at which to landmark the data}

\item{times_pred}{Times at which to predict the survival prob/AUC}

\item{M}{Number of dimensions to use in mFPCA}

\item{uniExpansions}{Univariate expansions to use in MFPCA}

\item{age}{Variable to perform alternative de-meaning with.}

\item{n_folds}{Number of folds for cross validation.}

\item{seed}{Seed for random number generation}

\item{mdat}{A multiFunData object containing longitudinal covariates}
}
\description{
Cross validate MFPCCox procedure
}
