% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/univariateExpansions.R
\name{fpcaFunction}
\alias{fpcaFunction}
\title{Calculate a linear combination of a functional principal component basis on
one-dimensional domains}
\usage{
fpcaFunction(scores, argvals = functions@argvals, functions)
}
\arguments{
\item{scores}{A matrix of dimension \code{N x K}, representing the \code{K}
scores (coefficients) for each of the \code{N} observations.}

\item{argvals}{A list containing a vector of x-values. Defaults to
\code{functions@argvals}.}

\item{functions}{A \code{funData} object, representing the FPCA basis.}
}
\value{
An object of class \code{funData} with \code{N} observations on
  \code{argvals}, corresponding to the linear combination of the functional
  principal components.
}
\description{
This function calculates  a linear combination of functional principal
component basis functions on one-dimensional domains.
}
\seealso{
\code{\link{univExpansion}}, \code{\link{fpcaBasis}}
}

