% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/univariateExpansions.R
\name{fpcaFunction}
\alias{fpcaFunction}
\title{Calculate a linear combination of a functional principal component basis on
one-dimensional domains}
\usage{
fpcaFunction(scores, xVal, functions)
}
\arguments{
\item{scores}{A matrix of dimension \eqn{N x K}, representing the \eqn{K}
scores (coefficients) for each observation \eqn{i = 1, \ldots, N}.}

\item{xVal}{A list containing a vector of x-values.}

\item{functions}{A \code{funData} object, representing the FPCA basis.}
}
\value{
An object of class \code{funData} with \eqn{N} observations on
  \code{xVal}, corresponding to the linear combination of the functional
  principal components.
}
\description{
This function calculates  a linear combination of functional principal
component basis functions on one-dimensional domains.
}
\seealso{
univExpansion
}

