% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_rcvlist}
\alias{plot_rcvlist}
\title{Plot performance measures of models}
\usage{
plot_rcvlist(
  x,
  legend = "",
  legend_pos = NULL,
  title = NULL,
  main_title = "",
  sub_title = "",
  col = NULL,
  lty = NULL,
  ...
)
}
\arguments{
\item{x}{A list containing "rcv_mfpccox" objects}

\item{legend}{A list of length(x) containing a description of each object in x.}

\item{legend_pos}{Legend position. See plot()}

\item{title}{Title for legend.}

\item{main_title}{Main title for plot.}

\item{sub_title}{Subtitle for plot.}

\item{col}{A vector of length(x) specifying which color to use for each object in x. Default uses qual pallette from brewer.pal.}

\item{...}{Further parameters to plot()}
}
\description{
Can plot output of rcv_mfpccox() functions.
}
\section{Functions}{
\itemize{
\item \code{plot_rcvlist()}: Plot prediction performance of rcv_mfpccox() output

}}
