% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_surv.R
\name{predict_surv}
\alias{predict_surv}
\title{Predict survival probabilities for new patients (step 3)}
\usage{
predict_surv(
  step2,
  times_pred,
  reg_baseline = FALSE,
  reg_long = TRUE,
  alpha = 1,
  accuracy_train = NULL,
  IPCW_vars = c("none", "all")
)
}
\arguments{
\item{step2}{output of get_mscores}

\item{times_pred}{Times at which prediction is required. Default = landmark_time (useless).}

\item{reg_baseline}{Should baseline covariates be regularized?}

\item{reg_long}{Should longitudinal covariates be regularized?}

\item{alpha}{Regularization scaling. Alpha = 1 (lasso), alpha = 0 (ridge), between 0 and 1 = elastic net. Default = 1 (lasso)}

\item{accuracy_train}{Should accuracy measures (AUC/Brier score) be calculated on the training set?}

\item{IPCW_vars}{Character vector indicating which variables to use for IPCW.
Extra options are "all" (all baseline variables) and "none" (don't perform IPCW). Default is no IPCW ("none").}
}
\description{
Predict survival for prediction data
}
\keyword{internal}
