% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFPCAfit_methods.R
\name{print.MFPCAfit}
\alias{print.MFPCAfit}
\title{Print the results of a Multivariate Functional Principal Component Analysis}
\usage{
\method{print}{MFPCAfit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{MFPCAfit}, usually returned by a
call to \link{MFPCA}.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
No return value, called for side effects
}
\description{
A \code{print} function for class \code{MFPCAfit}.
}
