% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/univDecomp.R
\name{splineBasis2D}
\alias{splineBasis2D}
\title{Calculate an unpenalized spline basis representation for functional data on
two-dimensional domains}
\usage{
splineBasis2D(funDataObject, bs, m, k)
}
\arguments{
\item{funDataObject}{An object of class \code{\link[funData]{funData}}
containing the observed functional data samples and for which the bases
representation is calculated.}

\item{bs}{An array (or a single character), the type of basis functions to be
used. Please refer to \code{\link[mgcv]{te}} for a list of possible basis
functions.}

\item{m}{An array (or a single character), the order of the spline basis. See
\code{\link[mgcv]{s}} for details.}

\item{k}{An array (or a single character), the number of basis functions
  used.  See  \code{\link[mgcv]{s}} for details.}
}
\value{
\item{scores}{A matrix of weights with dimension \code{N x K},
  reflecting the weights for each basis function in each observation, where
  \code{K} is the total number of basis functions used.}
   \item{B}{A matrix
  containing the scalar product of all pairs of basis functions.}
  \item{ortho}{Logical, set to \code{FALSE}, as basis functions are not
  orthogonal.} \item{functions}{\code{NULL}, as basis functions are known.}
}
\description{
This function calculates an unpenalized tensor product spline basis
representation for functional data on two-dimensional domains based on the
\link[mgcv]{gam} function in the \pkg{mgcv} package.
}

