% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/univDecomp.R
\name{splineBasis2Dpen}
\alias{splineBasis2Dpen}
\title{Calculate a penalized spline basis representation for functional data on
two-dimensional domains}
\usage{
splineBasis2Dpen(funDataObject, bs = "ps", m = NA, k = -1,
  parallel = FALSE)
}
\arguments{
\item{funDataObject}{An object of class \code{\link[funData]{funData}}
containing the observed functional data samples and for which the basis
representation is calculated.}

\item{bs}{A vector of character strings (or a single character string),
specifying the type of basis functions to be used. Defaults to "ps"
(P-spline functions). Please refer to \code{\link[mgcv]{te}} for a list of
possible basis functions.}

\item{m}{A numeric vector (or a single number), the order of the spline
basis. Defaults to NA, i.e. the order is chosen automatically.  See
\code{\link[mgcv]{s}} for details.}

\item{k}{An numeric vector (or a single number), the number of basis
functions used.  Defaults to -1, i.e. the number of basis functions is
chosen automatically.   See  \code{\link[mgcv]{s}} for details.}

\item{parallel}{Logical. If \code{TRUE}, the coefficients for the basis
  functions are calculated in parallel. The implementation is based on the
  \code{\link[foreach]{foreach}} function and requires a parallel backend
  that must be registered before. See \code{\link[foreach]{foreach}} for
  details.}
}
\value{
\item{scores}{A matrix of scores (coefficients) with dimension
  \code{N x K}, reflecting the weights for each basis function in each
  observation, where \code{K} is the total number of basis functions used.}
  \item{B}{A matrix containing the scalar product of all pairs of basis
  functions.} \item{ortho}{Logical, set to \code{FALSE}, as basis functions
  are not orthonormal.} \item{functions}{\code{NULL}, as basis functions are
  known.} \item{settings}{A list with entries \code{bs},
  \code{m} and \code{k}, giving the actual parameters used for generating the
  spline basis functions.}
}
\description{
This function calculates a penalized tensor product spline basis
representation for functional data on two-dimensional domains based on the
\link[mgcv]{bam} function in the \pkg{mgcv} package (for large GAMs).
}
\seealso{
univDecomp
}

