% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/univariateExpansions.R
\name{splineBasis2Dpen}
\alias{splineBasis2Dpen}
\title{Calculate a penalized spline basis representation for functional data on
two-dimensional domains}
\usage{
splineBasis2Dpen(funDataObject, bs, m, k)
}
\arguments{
\item{funDataObject}{An object of class \code{\link[funData]{funData}}
containing the observed functional data samples and for which the bases
representation is calculated.}

\item{bs}{An array (or a single character), the type of basis functions to be
used. Please refer to \code{link[mgcv]{te}} for a list of possible basis
functions.}

\item{m}{An array (or a single character), the order of the spline basis. See
\code{link[mgcv]{s}} for details.}

\item{k}{An array (or a single character), the number of basis functions
  used.  See  \code{link[mgcv]{s}} for details.}
}
\value{
scores A matrix of weights with dimension \code{N x k}, reflecting
  the weights for each basis function in each observation.

function A \code{\link[funData]{funData}} object representing the
  basis functions.

basisLong A matrix representing the basis functions in vectorized
  form.
}
\description{
This function calculates a penalized tensor product spline basis
representation for functional data on two-dimensional domains based on the
\link[mgcv]{bam} function in the \pkg{mgcv} package (for large GAMs).
}
\keyword{internal}

