% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFPCAfit_methods.R
\name{summary.MFPCAfit}
\alias{summary.MFPCAfit}
\title{Summarize a Multivariate Functional Principal Component Analysis}
\usage{
\method{summary}{MFPCAfit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{MFPCAfit}, usually returned by a
call to \link{MFPCA}.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
A \code{summary} method for class \code{MFPCAfit}
}
