% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UMPCA.R
\name{ttv}
\alias{ttv}
\title{Tensor times vector calculation}
\usage{
ttv(A, v, dim)
}
\arguments{
\item{A}{An array}

\item{v}{A list of the same length as dim}

\item{dim}{A vector specifying the dimensions for the multiplication}
}
\value{
An array, the result of the multiplication
}
\description{
Functionality adapted from MATLAB tensor toolbox
}

