% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/univariateExpansions.R
\name{umpcaFunction}
\alias{umpcaFunction}
\title{Calculate a linear combination of a UMPCA basis on two-dimensional domains}
\usage{
umpcaFunction(scores, argvals, functions)
}
\arguments{
\item{scores}{A matrix of dimension \eqn{N x K}, representing the \eqn{K}
scores (coefficients) for each observation \eqn{i = 1, \ldots, N}.}

\item{argvals}{A list containing a vector of x-values.}

\item{functions}{A \code{funData} object, representing the UMPCA basis.}
}
\value{
An object of class \code{funData} with \eqn{N} observations on
  \code{argvals}, corresponding to the linear combination of the functional
  principal components.
}
\description{
This function calculates a linear combination of basis functions based on
uncorrelated multilinear principal component analysis (UMPCA) for data on
two-dimensional domains.
}
\seealso{
univExpansion
}

